完善战斗逻辑中，攻击与被攻击的伤害计算逻辑

+++ src/main/java/cn/yans/onegame/entity/BaseCard.java
@@ -27,7 +27,7 @@ public class BaseCard {
     /**
      * 基础数值
      */
-    private double value;
+    private Long value;
     /**
      * 数值类型
      */
